35coreos-multipath: propagate config on first boot
In #829, we changed multipath config propagation to happen on subsequent
boots to handle upgrading nodes. Back then, we didn't yet support first
boot multipath, so not propagating on first boot didn't matter anyway.

But with #1011 we now support multipath on first boot. So we need to
adapt this so we always propagate, including on first boot.