workflows: add job to check for unexpected changes in promotion diff
When we get a promotion PR in the next/testing/stable branches, compare
the new code with a checkout of the commit hash from the PR title and
annotate the differences.  For manifest.yaml, compare with the base
branch instead, since the branch we're promoting from will have a
different config.