denylist: snooze the ext.config.firewall.iptables test on next*
We merged an update to the test that checks for iptables-nft on
next/next-devel but we haven't switched to iptables-nft there yet
because we are coupling that with the rebase to F36.

I imagine what we'll do is modify the test to be conditional on
Fedora 36 rather than the stream but we have to wait to make that
change until we actually switch rawhide/branched over as well,
which will most likely happen after https://github.com/coreos/rpm-ostree/pull/3439
lands.