overrides: drop kernel pin
This allows us to get the latest kernel-5.16.12-200.fc35. Moving to
a kernel newer than 5.16.11 picks up the fix fo CVE-2022-0847. We're
able to do this because the Fedora kernel maintainers agreed to again
pick up a revert that allows us to not regress on some AWS instance
types (https://github.com/coreos/fedora-coreos-tracker/issues/1066).

Closes https://github.com/coreos/fedora-coreos-tracker/issues/1118