manifests/fedora-coreos-base: mask systemd-network-generator.service
systemd-network-generator fails to run because of SELinux. The existing
bug for this problem [1] has been open for months. Let's mask the service
and move on. We do want the generator to run in the future for things
like creating udev rules [2] but it's not 100% necessary for now.

We're making this change now because we're about to switch `next` over
to Fedora 36 where we can't just freeze on an older version of systemd
and we don't want to mask the tests any longer because we want the test
coverage.

[1] https://bugzilla.redhat.com/show_bug.cgi?id=2037047
[2] https://github.com/systemd/systemd/pull/21766