tests: limit host bound package layer to specific streams
It's easy for some streams to not be able to package layer the vim
package because of the vim RPMs that are already installed in FCOS.
This can happen often on streams that aren't stable Fedora releases.

Let's just limit this part of the test to streams that are usually
based on Fedora stable versions.