manifests: fedora-coreos-base: add containernetworking-plugins explicitly
In order to continue to support CNI networks in podman, which is what
any existing Fedora <= 35 system is using, we need to continue to ship
containernetworking-plugins.

It was dropped in [1]. Traditional yum based systems won't uninstall the
package on an upgrade, but OSTrees are "baked fresh" every time and
won't retain it because it's no longer a dependency. We need to name it
explicitly if we want upgrading FCOS users to be able to start their
containers.

Additionally, since both `containernetworking-plugins` and `netavark`
provide `container-network-stack` if we want them both installed we
need to name them both in our manifests.

[1] https://src.fedoraproject.org/rpms/podman/c/06c601234feac4f5da7b6e1d92430d62f556f3b9