overlay: 40ignition-ostree: make ignition-ostree-transposefs tolerant to debug kernels
It appears that on debug kernels the "/dev/disk/by-label/$label" device *can* not
appear at all (some sort of race condition) and thus calling `realpath` on it will
fail. Let's just make the call to `realpath` not be fatal so we can workaround this
issue that has been around in the kernel for some time.

Closes https://github.com/coreos/fedora-coreos-tracker/issues/1092