manifest: explicitly list containerd
This is already in the OS as a dependency of moby-engine, but Typhoon is
now also relying on this and likely other people deploying Kubernetes on
top of FCOS.

Unlike cri-o, the versioning of containerd wrt CRI is less tight, which
makes baking it into the host easier.

Things may change here in the future, and it's likely we will recommend
cri-o as an alternative runtime for k8s eventually. But for now, let's
at least be more explicit that we're shipping containerd.

For more information, see:
https://github.com/coreos/fedora-coreos-tracker/issues/767
https://github.com/poseidon/typhoon/issues/899#issuecomment-1010638811