make kdump work on first boot
Up until this point people have had to reboot once after setting up
kdump so that if a crash happens the crash kernel doesn't get started
with the ignition.firstboot kernel argument. If we add `ignition.firstboot`
to the list of arguments in `KDUMP_COMMANDLINE_REMOVE=` in
`/etc/sysconfig/kdump` then we can set it up on first boot without issues.

Here we also switch the kdump.crash test to set up kdump via Ignition
to prove out the change works as desired.