ci/continuous: New container that pulls from continuous COPR
History goes in cycles.  One of the key original raison d'être of
ostree was having people consume git main/master of projects.  But...
because RPM doesn't make it easy to revert things, we really
end up going only at the same speed as the rest of Fedora sadly.

Anyways, we've talked about this a few times in FCOS; xref
https://github.com/coreos/fedora-coreos-tracker/issues/910

This adds a container image that is built *as* a container layering
on top of the base image, using the recently landed
https://github.com/coreos/rpm-ostree/pull/3605/

My plan is to add build automation (in the pipeline? or OCP Prow?)
that sticks this at e.g. `quay.io/coreos-assembler/fcos:35-continous`.