ci/overrides: don't load sack if overrides empty
Right now, `next-devel` and `branched` are inactive. But to reduce
churn, the `remove-graduated-overrides` GitHub Action is left running on
those branches since it's harmless. However, the repos on inactive
branches might not be valid, causing workflow failures.

But anyway, there are no active overrides on these branches, so we can
work around this by exiting earlier.

I think a more appropriate fix for this is to query whether `next-devel`
is open or closed (though we'd need something similar for `branched`
too), but we do also want this optimization anyway.