workflows: add scheduled job to sync openshift/os submodule
We've been hand-maintaining the fedora-coreos-config submodule in
openshift/os, but updating it is usually completely mechanical.  Add a
workflow that maintains an openshift/os PR with the latest changes,
updating it daily or on demand.  We can merge the PR whenever we'd like,
and the job will open a new one once there are additional changes.

Omit CoreOS Bot changes from the log summary, since those involve
lockfiles and overrides that don't apply to RHCOS.  Likewise, if the only
changes are from CoreOS Bot, skip opening a PR.

When the workflow is manually invoked, allow overriding the PR branch.
This can be used to create a one-off PR that won't be maintained by the
bot afterward, for cases where a sync needs complementary changes
elsewhere in openshift/os.

Ideally the job would run inside the openshift/os repo, but GitHub Actions
isn't available there.