workflows: add stamp file in openshift/os to ensure we push rebases
The create-pull-request action currently only force-pushes branches if
their diff has changed.  However, if we've meanwhile landed a parallel
fedora-coreos-config update PR via a different branch, the diff might
be the same but the base commit will be different.  We need to push the
branch in this case too, or else the PR will be left unmergeable due to
conflicts.  As a workaround, write the base commit ID into a stamp file
in the root of openshift/os, so the diff will change whenever the base
does.