s390x: drop 's390utils-base'
This package provides the `genprotimg` tool, but also depends on `perl-*` rpms,
which we don't want to be part of the image. As long as IBM Secure Execution
is targeted for RHCOS we can drop its support in FCOS for now.
RHCOS comes with a lot of `perl-*` rpms, so there `s390utils-base` wouldn't
bring new unwanted dependencies. For custom FCOS builds it's easier to:
```
$ mkdir -p overrides/rootfs/usr/bin/
$ mkdir -p overrides/rootfs/usr/share/s390-tools/genprotimg
$ cp /path/to/genprotimg overrides/rootfs/usr/bin/
$ cp /path/to/stage3a.bin overrides/rootfs/usr/share/s390-tools/genprotimg/
$ cp /path/to/stage3b_reloc.bin overrides/rootfs/usr/share/s390-tools/genprotimg/
```

https://github.com/coreos/fedora-coreos-tracker/issues/1217

Signed-off-by: Nikita Dubrovskii <nikita@linux.ibm.com>