manifests/group: drop well-known static ID for "tss" group
This drops a redundant override for the "tss" group.
The group is created by the "tpm2-tss" package and has a static
group ID (GID) allocated in the `uidgid` table.
Out of an abundance of caution, it also temporarily removes a
buggy sysusers fragment which does not match Fedora allocation.

Ref: https://bugzilla.redhat.com/show_bug.cgi?id=2103683