tests: set min memory for chrony dhcp propagation
RHCOS 4.10 builds were failing because the qemu instance was running out
of memory. The test was failing as exit code 125 from the dockerfile
build but with exit code 137 when run manually. It was only running out
of memory downloading the repodata on the Fedora 35 container. The
Fedora 36 container did not have any issues.

Fixes: https://github.com/openshift/os/issues/899