manifests/group: fix ID for "input" group
This fixes the static GID for the `input` group, properly using the
Fedora-wide reserved ID `104`.
This group is mostly used as the owner of ephemeral entries under
`/dev`, thus it should be safe to update (upon booting into the new
deployment, devices will get created from scratch with the new GID).

Ref: https://pagure.io/setup/blob/108f62338e57652bb3a37e9028572704f3f28c29/f/uidgid#_113