Handle dm-verity-protected filesystems in s390x secex path
Instead of shipping the rootfs and bootfs using LUKS, we're pivoting to
verifying integrity using dm-verity instead.

Add `veritysetup` to the OS. Teach `transposefs` to transfer the
filesystem over dm-verity. Delete the root and boot hash partitions.
Delete some of the bits we previously added for LUKS enablement.

For more information, see:
https://github.com/coreos/coreos-assembler/pull/3039