Add fedora-rawhide-nodebug-kernel repo to fedora-rawhide.repo
We are choosing to use only nodebug kernels in Fedora CoreOS
for our testing. We've seen too many issues where an issue either
shows up only on the debug kernels OR a debug kernel gives us a
false positive or negative when trying to root cause a failure.
Thus we add the fedora-rawhide-nodebug-kernel repo here so that
we can use it in the rawhide manfest. This uses the nodebug kernel
repo [1].

[1] https://fedoraproject.org/wiki/RawhideKernelNodebug