manifests: drop explicit add of kernel-{core,modules}
kernel-{core,modules} get pulled in by the `kernel` meta package.
We previously named these explicitly because we saw some issues where
the `kernel-debug-core` RPM was getting chosen instead (see )

Let's solve the problem a different way by just naming `kernel` like we
did in the past, but include `kernel-debug-core` in the list of packages
we explicitly exclude. This has the benefit of making it simpler to run
kernel regression git bisect with the kernel rpm generated from
`make binrpm-pkg` in upstream kernle git, which just creates a single
RPM named `kernel`.

This has the benefits of simplifying our steps for how to build/test
an RPM from upstream kernel git [1] (permalink for example [2]) by
allowing us to not have to edit any manifest files as part of the process.

[1] https://github.com/coreos/fedora-coreos-tracker/blob/main/docs/fedora-coreos-kernel-bisect.md
[1] https://github.com/coreos/fedora-coreos-tracker/blob/e0a4c4fbec14248b4771ce30e0639996a2100153/docs/fedora-coreos-kernel-bisect.md#doing-a-build-with-cosa