tests: Switch one test to use platform-independent + YAML
Since these two PRs

- https://github.com/coreos/coreos-assembler/pull/3064/commits/242e7149b8d57b5837238366bfcd70bd99b455a5
- https://github.com/coreos/coreos-assembler/pull/3059/commits/302130ec171d3d501310c40bd28db49c21459519

we support YAML for kola metadata, and there's a new `platform-independent`
tag.

Use both in a test I added recently to demonstrate usage, and
this also serves to cross-check that things work.

Motivation: I think semantic tags and inline comments are far
superior to duplicating the test metadata and I don't want to
have to do that again in a test I write.