overrides: fast-track kernel-5.19.12-200.fc36
The revert of the problematic commit that caused the hung reboots
in our `coreos.boot-mirror` tests landed upstream in the 6.0 development
cycle and was backported to 5.19.12. This means the Fedora kernel
maintainer has dropped his manual revert of the commit and also that
rawhide should no longer see the issue (since it's already on 6.0 rc
kernels.

Closes https://github.com/coreos/fedora-coreos-tracker/issues/1282