Document new-style console configuration
Mainly this affects the emergency shell docs.  Also add a section to the
kargs page in case a user goes directly to that doc.

Delete the "temporary tweak" section, since it doesn't make sense now.
We previously needed it for removing unwanted console arguments so the
booted console would be usable.  Now, with per-platform defaults, the user
will likely want to add consoles rather than remove them, but in that case
GRUB won't show a menu on the desired console in the first place.