tests/kola: remove date conditional from files.console-config test
First, the dates mentioned in [1] have passed so we no longer need
this check. Second, checking the date inspecting the value from $VERSION
was actually not quite correct because `stable` lags behind so the
actual date in its $VERSION string is the date from the timestamp of
the yumrepos (which was two weeks in the past). This means the `stable`
threshold=20221209 isn't checking correctly (the current `stable` build
we are trying to build today has a version string of `37.20221127.3.0`).

Just dropping the check solves both of these problems. Once merged we'll
forward port this to the `stable` branch too.

[1] https://lists.fedoraproject.org/archives/list/coreos-status@lists.fedoraproject.org/message/GHLXX4MXNHUEAXQLK6BZN45IQYHRVQB4/