overlay.d: only run coreos-enable-network on first boot
I noticed recently that Azure is bringing up networking in the initramfs
on every boot [1]. This behavior was originally introduced in c1cf903. After
some discussion with @jlebon and @bgilbert we are going to convert the
coreos-enable-network.service to only be pulled in and enalbed on the
Ignition boot so that it doesn't run on subsequent boots.

[1] https://github.com/coreos/fedora-coreos-tracker/issues/1383#issuecomment-1384526420