overlay.d: create new 25-azure-udev-rules overlay
We previously had the 68-azure-sriov-nm-unmanaged.rules but we
figured out we also want those rules in the initramfs. This
commit makes a new overlay.d entry and groups together a new
dracut module along with the udev rules in that overlay.d directory.

Fixes https://github.com/coreos/fedora-coreos-tracker/issues/1383