tests: set rd.neednet=1 in bootif net propagation test
Normally on qemu networking isn't brought up in the initramfs because
it's not needed (the Ignition config is passed via fw_cfg, not network).
This will force networking to be brought up to verify that bringing up
NetworkManager doesn't materially change things.

One case where it did materially change things and we needed to adapt
is described in https://github.com/coreos/fedora-coreos-tracker/issues/1385