coreos-livepxe-rootfs: print initramfs version on error
If the rootfs doesn't match the initramfs, let's include the version of
the initramfs in the error message. This version string also appears in
e.g. the boot menu and initramfs welcome message, but having it be
directly part of the final error message is clearer and easier.

The only caveat to this is that we don't have easy access to the OSTree
version in RHCOS 8, since it has an ancient dracut which doesn't have
https://github.com/dracutdevs/dracut/pull/655. But it should work fine
on el9.

Related: https://github.com/coreos/fedora-coreos-tracker/issues/1390