coreos-livepxe-rootfs: use `curl --fail`
Currently, if the server returns a 404 when fetching the rootfs, we
still try to hash the server's HTML response and compare it, which
leads to a confusing error message about mismatched hashes.

Do a single `HEAD` after establishing server connectivity, but before
the pipeline. Use `--fail` so that `curl` exits with a nonzero code and
clearly prints out the erroring HTTP code it got. Also use `--fail`
during the pipeline invocation for the same reason.

Closes: https://github.com/coreos/fedora-coreos-tracker/issues/1390