tests/kola: test spec 3.4.0 LUKS discard and openOptions fields
Rather than adding a dedicated test for relatively obscure options,
change the rootfs test to enable them, check that it does so, and check
that the existing LUKS /var/log test doesn't do so.  In each case, check
both the first and second boots.

The rootfs is the important case, since its mount flow doesn't read
/etc/crypttab.