workflow: don't remove graduated overrides on next-devel when disabled
The `next-devel` stream uses the -next repos which recently became
invalid since it still assumes f37. The graduation workflow has been
failing since then and I've been receiving notifications each time it
runs.

Since `next-devel` is unmaintained when it's disabled, it doesn't
make sense to run this workflow there (e.g. there's currently a stale
NetworkManager pin there).

Rework the job so that the matrix is built dynamically, and we only
include `next-devel` in the matrix when it's enabled.