coreos-platform-chrony-config: use `ptp_kvm` when available on QEMU
KVM supports a `ptp_kvm` kernel module which allows the guest to query
the host clock for synchronization purposes.

Configure chrony to make use of it if it's available.

Don't disable the default pool; it'll be used by chrony to determine
whether the host clock is accurate or a falseticker.

Don't disable `PEERNTP` as we do on other platforms since we
historically haven't done so on QEMU and it's possible users have been
using their own DHCP servers (e.g. dnsmasq) to feed NTP configuration.

Closes: https://github.com/coreos/fedora-coreos-tracker/issues/1433

Co-authored-by: Jonathan Lebon <jonathan@jlebon.com>