Revert "overlay/15fcos: upgrade bootloader on aarch64 at boot"
This reverts commit 8ce6fd6827b06e3c462ddc8c18d2fe1bdc7f5e86.

The testing-devel promotion with this has already happened in
https://github.com/coreos/fedora-coreos-config/pull/2315
and we have already shipped a `next` with the unit, so we can
drop it now (before we execute the `next-devel`->`next` promotion)
to prevent it from shipping in more than one release per stream.

We can do this because we have barriers.

Full context in https://github.com/coreos/fedora-coreos-tracker/issues/1441