40ignition-ostree: run ignition-ostree-growfs before sysroot mount
Prep for automatic XFS reprovisioning.

The only way to know for sure whether the rootfs should be reprovisioned
is analyzing it after the filesystem was grown. We could do calculations
beforehand, but it'd get complex having to analyze the partition table.

Anyway, the partition growing and e.g. LUKS container resizing need
to happen before automatic reprovisioning and ignition-ostree-growfs
already knows how to do that.