overlay.d/05core: don't omit dracut modules we don't ship anyway
We don't ship rngd, busybox, pcscd, or bluetooth, so there's no point in
listing them as modules to omit; the modules already know they can't be
installed.

This allows users to more easily include them in the initrd if they
do decide to overlay those packages and want them in the initrd (by
enabling local initramfs regeneration). Another use case is kata-
containers, which creates the guest OS using bits from the host and
wants to be able to include busybox for debugging purposes.

Closes: https://github.com/coreos/fedora-coreos-tracker/issues/1450