overlay.d: Add/Update udev rules for GCP
 - This PR injects/updates the udev rules from google-guest-config package into
our overlay (GoogleCloudPlatform/guest-configs@18fbc05).

 - It works as an workaround for the BZ #2176212. The longer term solution
is to add these packages as part of the Fedora/RHEL itself, so we won't need to
maintain it, as we can entirely drop it.

For this matter the BZ https://bugzilla.redhat.com/show_bug.cgi?id=2182865
was opened requesting a subpackage of google-compute-engine-guest-configs
containing only what is needed.

For more info:

https://bugzilla.redhat.com/show_bug.cgi?id=2176212
https://github.com/GoogleCloudPlatform/guest-configs/tree/master/src/lib/udev

Signed-off-by: Renata Ravanelli <rravanel@redhat.com>