coreos-teardown-initramfs: teardown after NM has exited
We haven't seen this before, but recently there was
a case where in RHCOS based on RHEL 9 we saw the coreos-teardown-initramfs
happening before the nm-initrd.service had been stopped
during switch root. We need to guarantee that NetworkManager
is stopped before we start the teardown.

Seen in https://issues.redhat.com/browse/OCPBUGS-11052