tests/kola: bump `minDisk` in autosave-xfs tests
The actual agcount on a 64G disk depends on the agsize baked in the
image, which in turn depends on how large the rootfs was calculated to
be (see `estimate-commit-disk-size` in cosa).

Since the starting rootfs size will differ on different arches (notably,
be larger than x86_64) and minor variations are greatly amplified, the
agcount may be lower than expected for a 64G disk and in turn the test
can fail. Even on x86_64, the rootfs size can change e.g. in CI builds
where we bake non-stripped binaries.

Bump the requested disk size to 100G in the test to make it more
foolproof.

Closes: https://github.com/coreos/fedora-coreos-tracker/issues/1458