Send system journal logs to console
When debugging early boot issues, especially ones that leave the host
without functional networking, it's difficult to debug what broke. This
change makes the default to send the systemd journal logs to the console
on x86_64.  We attempted to do this in openshift with a machineconfig,
and this broke arm64 and this config was offered as an alternative.