overlay.d/05core: add ifname= karg udev rule propagation
This adds to coreos-teardown-initramfs.{service,sh} the ability to
propagate udev rules that were generated by ifname-genrules.sh [1]
to the real root of the machine to allow for an ephemeral ifname=
karg on the first boot to have persistent configuration effect (similar
to our other networking kernel arguments).

This also adds tests for both the firstboot and everyboot case for
the ifname= karg.

Fixes https://github.com/coreos/fedora-coreos-tracker/issues/553

[1] https://github.com/dracutdevs/dracut/blob/5c3d0a96473ac339fa2d1b25213b8f301c1cfd0d/modules.d/40network/ifname-genrules.sh