denylist: deny reprovision tests on ppc64le/rawhide
They are failing right now. See
https://github.com/coreos/fedora-coreos-tracker/issues/1489

Note that we can't really deny these tests without hitting a bug
(feature?) in that if all tests in a kola run are denylisted kola
will exit with an error. Therefore we add the reprovision label to
a random test (`ext.config.boot.grub2-install`) so we'll have at
least one test run.