rootless-pasta-networking: Verify that mtu got set
Change the way we were testing the setting we were making
on the command line did make it into the networking of
the container we created. Now we verify the mtu got set to 1500.
Earlier we assigned 8.8.8.8 as default gateway in container but
no matching address on the same subnet was configured.
Ref: https://github.com/coreos/fedora-coreos-tracker/issues/1509#issuecomment-1608831096
Also, unpinned the passt package since the test should pass against
any version of the package