Drop modularity support/tests
In Fedora 39 modularity is going away [1]. Let's drop the repos for F39+
and also drop the test that we have. We do still want to test installing
crio works so let's add crio to the ext.config.extensions.package test.

We also drop the repos as input to the build here. They were never used
for actual RPM content anyway.

[1] https://fedoraproject.org/wiki/Changes/RetireModularity