initramfs: Remove manual sysroot mounting
Our dependency graph for firstboot is today kind of broken
because we only implicitly generate `sysroot.mount` as a side
effect of executing `ignition-ostree-mount-sysroot.sh`
(which is part of `ignition-ostree-mount-{firstboot,subsequent}-sysroot.service`)s.

As of relatively recently, we switched to auto-synthesizing the `root=`
karg after ignition (in `coroes-boot-edit.service`).  This means
that the systemd-fstab-generator will always win on subsequent boots,
so the systemd unit `-subsequent-sysroot.service` is dead code.

On firstboot, let's generate `sysroot.mount` out of the diskful
generator, which will ensure it always exists and fix our dependency
graph.

This should improve compatibility with newer systemd, and just
make everything saner too.