tests/manual: add script to launch a Hyper-V runner in AWS
Most of the script is EC2 boilerplate to get a publicly routable IP
address.  Unfortunately our test accounts have a pruner that deletes the
default VPC, so we can't just use that.  For simplicity, create all our
own infrastructure, and don't clean it up because the pruner will get it.

Most of the rest of the script is boilerplate for connecting to a Windows
instance.  The only things that are Hyper-V specific are running on a
bare-metal instance and installing Hyper-V.