tutorials: drop secondary "local files" path; make it primary
In e96677c a secondary path in the tutorial was introduced where a user
can optionally do things a different way. In my experience this isn't
a good user experience for someone who is learning. The best thing is
for the lab content to be executed one way and not have a bunch of
conditionaly logic.

I think the original goal was to highlight some of the new features
of Butane. Let's just make those new local file embedding features
the main way it gets done and drop the alternative path.